@@ -21,15 +21,15 @@
   "dependencies": {
     "@algolia/client-search": "^4.23.2",
     "@crowdin/cli": "4.6.1",
-    "@docusaurus/core": "3.7.0",
-    "@docusaurus/plugin-client-redirects": "3.7.0",
-    "@docusaurus/plugin-content-docs": "3.7.0",
-    "@docusaurus/preset-classic": "3.7.0",
-    "@types/react": "^17.0.49",
+    "@docusaurus/core": "3.8.1",
+    "@docusaurus/plugin-client-redirects": "3.8.1",
+    "@docusaurus/plugin-content-docs": "3.8.1",
+    "@docusaurus/preset-classic": "3.8.1",
+    "@types/react": "^19.1.9",
     "clsx": "^1.2.1",
-    "prism-react-renderer": "^1.3.5",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
+    "prism-react-renderer": "^2.4.1",
+    "react": "^19.1.1",
+    "react-dom": "^19.1.1",
     "typescript": "^5.4.5",
     "webpack": "^5.92.0"
   },
